+++ src/router/index.ts
@@ -21,13 +21,13 @@ const routes: Array<RouteRecordRaw> = [
   {
     path: '/bookshelf',
     name: ROUTE_NAMES.bookshelf,
-    component: () => import('@/pages/BookshelfPage.vue'),
+    component: () => import('@/pages/Bookshelf/BookshelfPage.vue'),
   },
   {
     path: '/bookshelf/:id',
     name: ROUTE_NAMES.bookshelfItem,
     props: true,
-    component: () => import('@/pages/BookshelfItemPage.vue'),
+    component: () => import('@/pages/Bookshelf/BookshelfItemPage.vue'),
   },
   {
     path: '/my-nfts',