+++ CHANGELOG.md
@@ -5,11 +5,16 @@ Th format is based on [Keep a Changelog
 an this project adheres to [Semantic Versioning
 
 ## [Unreleased]
+#### Added
+- Network switcher component(doing nothing for now)
+
 #### Changed
 - About page redesign
 - Footer links
 - Mobile version redirects to metamask browser or to the page of its app
 - Connecting to metamask logic moved to useMetaMaskConnect hook
+- Main page redesign
+- Book page redesign
 
 ## [1.0.0] - 2022-11-30
 #### Added